mpdecimal: Optionally build unit tests
This requires adding a script to move the test data to the expected
location. Add the script to the list of allowed files. Executing all
tests requires downloading additional test data. It is not however
good practice to do that. Do not run these tests.

The tests duration is set excessively long to avoid timeouts in CI.